on_premise/spark_delta_lake/requirements.txt
notebook
findspark
pyspark
kafka-python